Lottie fixes and refactors.
