+++ Microsoft.Toolkit.Uwp.SampleApp/readme.md
@@ -38,7 +38,7 @@ You can define "interactive" values in this file. The value types can be:
 * String: You want the user to provide a text. The string is built like this @[Name:**String**:Default value:UpdateSourceTriggerId]
 * Slider: You want the user to provide a double value. The string is built like this @[Name:**Slider**:Default value:min-max]
 * DoubleSlider: Same as slider but with double values (0.01 precision)
-* RangeSelectorMin and RangeSelectorMax: You want the user to provide two values, usualy for max and min. The strings are built like this @[NameMin:RangeSelectorMin:Default min value:min:ID]@ and @[NameMax:RangeSelectorMax:Default max value:max:ID]@. The IDs need to match.
+* RangeSelectorMin and RangeSelectorMax: You want the user to provide two values, usualy for max and min. The strings are built like this @[NameMin:RangeSelectorMin:Default min value:min:ID] and @[NameMax:RangeSelectorMax:Default max value:max:ID]. The IDs need to match.
 * TimeSpan: You want the user to provide a duration. The string is built like this (all values in miliseconds) @[Name:**TimeSpan**:DefaultValue:min-max]
 * Enum: You want the user to provide a enum value. The string is built like this @[Name:**Enum**:EnumType.DefaultValue]
 * Brush: You want the user to select a color from a list. The string is built like this @[Name:**Brush**:Black]
